requirements.txt
pytest
selenium
allure-pytest
python-dotenv
click
webdriver-manager
pytest-html